https://www.youtube.com/watch?v=totdD9wDWqc
I have made a memory in visual studio code and both followed a video called Build a Memory Card Game Live! Beginners Projects Complete Guide by GeeksforGeeks and reworked it so that it would be from my style.

The link to the video is both in the assignment and here.